producer/requirements.txt
kafka-python==2.0.2
feedparser==6.0.10